my Fedora + SwayWM configuration files
Some of these are my own but most of them are the files i either got from the official repos or by other people and then i tweaked them to my prefrences. The reason to keep them here is to keep them organized. Feel free to copy any of them you like.